Refactor folder structure,
- Add regex field to stack
- Modify linting rules
- Add basic tests
