+++ src/utils/scriptExtensions.js
@@ -61,7 +61,7 @@ module.exports = {
             "102F8",
             "102F9",
             "102FA",
-            "102FB",
+            "102FB"
         ],
         Beng: [
             "0951",
@@ -90,7 +90,7 @@ module.exports = {
             "1CF5",
             "1CF6",
             "1CF7",
-            "A8F1",
+            "A8F1"
         ],
         Bopo: [
             "3001",
@@ -132,7 +132,7 @@ module.exports = {
             "FF62",
             "FF63",
             "FF64",
-            "FF65",
+            "FF65"
         ],
         Bugi: ["A9CF"],
         Buhd: ["1735", "1736"],
@@ -156,7 +156,7 @@ module.exports = {
             "1046",
             "1047",
             "1048",
-            "1049",
+            "1049"
         ],
         Copt: [
             "102E0",
@@ -186,7 +186,7 @@ module.exports = {
             "102F8",
             "102F9",
             "102FA",
-            "102FB",
+            "102FB"
         ],
         Cpmn: ["10100", "10101"],
         Cprt: [
@@ -246,7 +246,7 @@ module.exports = {
             "1013C",
             "1013D",
             "1013E",
-            "1013F",
+            "1013F"
         ],
         Cyrl: ["0483", "0484", "0485", "0486", "0487", "1DF8", "2E43", "A66F"],
         Deva: [
@@ -317,7 +317,7 @@ module.exports = {
             "A838",
             "A839",
             "A8F1",
-            "A8F3",
+            "A8F3"
         ],
         Dogr: [
             "0964",
@@ -341,7 +341,7 @@ module.exports = {
             "A836",
             "A837",
             "A838",
-            "A839",
+            "A839"
         ],
         Dupl: ["1BCA0", "1BCA1", "1BCA2", "1BCA3"],
         Geor: ["10FB"],
@@ -382,7 +382,7 @@ module.exports = {
             "1133C",
             "11FD0",
             "11FD1",
-            "11FD3",
+            "11FD3"
         ],
         Grek: ["0342", "0345", "1DC0", "1DC1"],
         Gujr: [
@@ -409,7 +409,7 @@ module.exports = {
             "A836",
             "A837",
             "A838",
-            "A839",
+            "A839"
         ],
         Guru: [
             "0951",
@@ -435,7 +435,7 @@ module.exports = {
             "A836",
             "A837",
             "A838",
-            "A839",
+            "A839"
         ],
         Hang: [
             "3001",
@@ -473,7 +473,7 @@ module.exports = {
             "FF62",
             "FF63",
             "FF64",
-            "FF65",
+            "FF65"
         ],
         Hani: [
             "3001",
@@ -763,7 +763,7 @@ module.exports = {
             "1D370",
             "1D371",
             "1F250",
-            "1F251",
+            "1F251"
         ],
         Hano: ["1735", "1736"],
         Hira: [
@@ -818,7 +818,7 @@ module.exports = {
             "FF65",
             "FF70",
             "FF9E",
-            "FF9F",
+            "FF9F"
         ],
         Java: ["A9CF"],
         Kali: ["A92E"],
@@ -874,7 +874,7 @@ module.exports = {
             "FF65",
             "FF70",
             "FF9E",
-            "FF9F",
+            "FF9F"
         ],
         Khoj: [
             "0AE6",
@@ -896,7 +896,7 @@ module.exports = {
             "A836",
             "A837",
             "A838",
-            "A839",
+            "A839"
         ],
         Knda: [
             "0951",
@@ -923,7 +923,7 @@ module.exports = {
             "A832",
             "A833",
             "A834",
-            "A835",
+            "A835"
         ],
         Kthi: [
             "0966",
@@ -945,7 +945,7 @@ module.exports = {
             "A836",
             "A837",
             "A838",
-            "A839",
+            "A839"
         ],
         Latn: [
             "0363",
@@ -976,7 +976,7 @@ module.exports = {
             "A705",
             "A706",
             "A707",
-            "A92E",
+            "A92E"
         ],
         Limb: ["0965"],
         Lina: [
@@ -1024,7 +1024,7 @@ module.exports = {
             "10130",
             "10131",
             "10132",
-            "10133",
+            "10133"
         ],
         Linb: [
             "10100",
@@ -1083,7 +1083,7 @@ module.exports = {
             "1013C",
             "1013D",
             "1013E",
-            "1013F",
+            "1013F"
         ],
         Mahj: [
             "0964",
@@ -1107,7 +1107,7 @@ module.exports = {
             "A836",
             "A837",
             "A838",
-            "A839",
+            "A839"
         ],
         Mand: ["0640"],
         Mani: ["0640", "10AF2"],
@@ -1122,7 +1122,7 @@ module.exports = {
             "A836",
             "A837",
             "A838",
-            "A839",
+            "A839"
         ],
         Mong: ["1802", "1803", "1805", "202F"],
         Mult: [
@@ -1135,7 +1135,7 @@ module.exports = {
             "0A6C",
             "0A6D",
             "0A6E",
-            "0A6F",
+            "0A6F"
         ],
         Mymr: [
             "1040",
@@ -1148,7 +1148,7 @@ module.exports = {
             "1047",
             "1048",
             "1049",
-            "A92E",
+            "A92E"
         ],
         Nand: [
             "0964",
@@ -1171,7 +1171,7 @@ module.exports = {
             "A832",
             "A833",
             "A834",
-            "A835",
+            "A835"
         ],
         Nkoo: ["060C", "061B", "061F", "FD3E", "FD3F"],
         Orya: ["0951", "0952", "0964", "0965", "1CDA", "1CF2"],
@@ -1193,7 +1193,7 @@ module.exports = {
             "A836",
             "A837",
             "A838",
-            "A839",
+            "A839"
         ],
         Sinh: ["0964", "0965"],
         Sogd: ["0640"],
@@ -1209,7 +1209,7 @@ module.exports = {
             "09EC",
             "09ED",
             "09EE",
-            "09EF",
+            "09EF"
         ],
         Syrc: [
             "060C",
@@ -1230,7 +1230,7 @@ module.exports = {
             "0655",
             "0670",
             "1DF8",
-            "1DFA",
+            "1DFA"
         ],
         Tagb: ["1735", "1736"],
         Takr: [
@@ -1245,7 +1245,7 @@ module.exports = {
             "A836",
             "A837",
             "A838",
-            "A839",
+            "A839"
         ],
         Tale: [
             "1040",
@@ -1257,7 +1257,7 @@ module.exports = {
             "1046",
             "1047",
             "1048",
-            "1049",
+            "1049"
         ],
         Taml: [
             "0951",
@@ -1286,7 +1286,7 @@ module.exports = {
             "1133C",
             "11FD0",
             "11FD1",
-            "11FD3",
+            "11FD3"
         ],
         Telu: ["0951", "0952", "0964", "0965", "1CDA", "1CF2"],
         Tglg: ["1735", "1736"],
@@ -1306,7 +1306,7 @@ module.exports = {
             "0668",
             "0669",
             "FDF2",
-            "FDFD",
+            "FDFD"
         ],
         Tirh: [
             "0951",
@@ -1323,7 +1323,7 @@ module.exports = {
             "A836",
             "A837",
             "A838",
-            "A839",
+            "A839"
         ],
         Yezi: [
             "060C",
@@ -1338,7 +1338,7 @@ module.exports = {
             "0666",
             "0667",
             "0668",
-            "0669",
+            "0669"
         ],
         Yiii: [
             "3001",
@@ -1366,7 +1366,7 @@ module.exports = {
             "FF62",
             "FF63",
             "FF64",
-            "FF65",
+            "FF65"
         ],
     },
 };
